Don't run pre-commit on windows

+++ bindings/python/.evergreen/test.sh
@@ -75,13 +75,16 @@ else
 fi
 
 
-# Only run once and with Python 3.8+
-createvirtualenv $BASE_PYTHON .venv
-python -m pip install certifi
-python -m pip install pre-commit
-pre-commit run --all-files
-deactivate
-rm -rf .venv
+# Don't run pre-commit on Windows
+if [ "$OS" != "Windows_NT" ];
+  # Only run once and with Python 3.8+
+  createvirtualenv $BASE_PYTHON .venv
+  python -m pip install certifi
+  python -m pip install pre-commit
+  pre-commit run --all-files
+  deactivate
+  rm -rf .venv
+fi
 
 for PYTHON_BINARY in "${PYTHONS[@]}"; do
     echo "Running test with python: $PYTHON_BINARY"
